backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
onnxruntime==1.17.0
numpy==1.26.4
pillow
opencv-python-headless
httpx==0.24.1
